+++ CHANGELOG.md
@@ -22,6 +22,11 @@
         + `TableWindowInfo`
             + Regex title recognition (with tests)
 
+* Feature/UI/Seat_Tracking/0.5.0
+
+    + Color range adjustments / UI refinements
+    + Added `Tables.ColorRanges`
+
 * Feature/UI/Seat_Tracking/0.4.8
 
     + Created `Right Seat View Controller` (with mirrored assets)